Powers switching
Testing on Spawn Wall Super power:
idea: Movement Recognition -> switching SuperPower Mode -> Spawn Wall on X Button Pressed

ERROR on WallSpowner script. testing and work in progress.
