+++ VR_Game1/Assets/Prefabs/Assets_Base/Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 637d89d016efcaa42b637a57d7babdcf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
