+++ VR_Game1/Assets/Prefabs/Assets_Base/VR Camera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f1bdebbbe1a46d8458310858a5f04c08
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
