+++ VR_Game1/Assets/Scripts/PowersManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65687cc5e62259b42acc9fb11879a83a
+guid: 0758f530cc9d6ac4bbc8acc7682392fc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
