+++ VR_Game1/Assets/Scripts/WallSpawner.cs
@@ -11,7 +11,8 @@ public class WallSpawner : MonoBehaviour
     public GameObject wall;
     public GameObject cam;
     public GameObject player;
-    public float spawnDistance;
+    //public float spawnDistance; testing 
+    float spawnDistance = 5.0f;
     private bool isWallBuilt = false;
     
     public List<GameObject> objects;
@@ -28,7 +29,7 @@ public class WallSpawner : MonoBehaviour
     {
         float dirX = cam.transform.forward.x;
         float dirY = player.transform.forward.y;
-        float dirZ = cam.transform.forward.z * spawnDistance;
+        float dirZ = cam.transform.forward.z * spawnDistance; 
         
         Vector3 playerPos = cam.transform.position;
         Vector3 playerDirection = new Vector3(dirX, dirY,dirZ);  //player.transform.forward; 
@@ -37,8 +38,9 @@ public class WallSpawner : MonoBehaviour
         if (isWallBuilt == false)
         {
             var cloneWall = Instantiate(wall, spawnPos, playerRotation);
-            Debug.Log("Coordinate Camera "+ cam.transform.forward.x + " -- " + player.transform.forward.y + " -- " + cam.transform.forward.z * spawnDistance);
-            Debug.Log("Muro. " + cloneWall.transform.position.x + " -- " + cloneWall.transform.position.y + " -- " + cloneWall.transform.position.z);
+            Debug.Log("Muro Spownato");
+           //Debug.Log("Coordinate Camera "+ cam.transform.forward.x + " -- " + player.transform.forward.y + " -- " + cam.transform.forward.z * spawnDistance);
+           //Debug.Log("Muro. " + cloneWall.transform.position.x + " -- " + cloneWall.transform.position.y + " -- " + cloneWall.transform.position.z);
             isWallBuilt = true;
             Destroy(cloneWall, 7.0f);
         }
